- Makefile、evallearnを指定してmakeした時に、EVAL_LEARNのシンボルを有効化してコンパイルするように。 	- Makefileのdevを追加。DEV_BRANCHのコンパイル用。 		- DEV_BRANCHシンボルの導入。supersortなどはここに入れておく。 	- Makefileのコード、整理した。 	- Clangでwarningが出てた箇所、いくつか修正。 	- USI_Hashの最大サイズ(MaxHashMB)、新しい上限サイズに書き換え忘れてた。修正。

+++ source/tt.cpp
@@ -261,7 +261,7 @@ void TranspositionTable::init_tt_per_thread()
 	ASSERT_LV3((clusterCountPerThread & 1) == 0);
 
 	// これを、自分が確保したglobalな置換表用メモリから切り分けて割当てる。
-	for (int i = 0; i < thread_size; ++i)
+	for (size_t i = 0; i < thread_size; ++i)
 	{
 		auto& tt = Threads[i]->tt;
 		tt.clusterCount = clusterCountPerThread;
